Move exceptions to emails/exceptions.py
